Extract common answer flag parsing
